moved serve models into api folder
