+++ src/maps/uiFlowRunHistory.ts
@@ -1,6 +1,6 @@
 
+import { UiFlowRunHistoryResponse } from '@/models/api/UiFlowRunHistoryResponse'
 import { UiFlowRunHistory } from '@/models/UiFlowRunHistory'
-import { UiFlowRunHistoryResponse } from '@/models/UiFlowRunHistoryResponse'
 import { MapFunction } from '@/services/Mapper'
 
 export const mapUiFlowRunHistoryResponseToUiFlowRunHistory: MapFunction<UiFlowRunHistoryResponse, UiFlowRunHistory> = function(source: UiFlowRunHistoryResponse): UiFlowRunHistory {
