--- src/models/DeploymentFlowRunRequest.ts
@@ -1,14 +1,0 @@
-import { StateRequest } from '@/models/StateRequest'
-import { SchemaValues } from '@/types/schemas'
-
-export type DeploymentFlowRunRequest = Partial<{
-  name: string | null,
-  parameters: SchemaValues,
-  idempotency_key: string | null,
-  context: unknown,
-  empirical_policy: unknown,
-  tags: string[] | null,
-  parent_task_run_id: string | null,
-  infrastructure_document_id: string | null,
-  state: Partial<StateRequest>,
-}>
