--- src/models/EmpiricalPolicyResponse.ts
@@ -1,6 +1,0 @@
-export type EmpiricalPolicyResponse = {
-  retries: number | null,
-  retry_delay: number | null,
-  max_retries: number | null,
-  retry_delay_seconds: number | null,
-}
