--- src/models/FlowResponse.ts
@@ -1,7 +1,0 @@
-export type FlowResponse = {
-  created: string,
-  id: string,
-  name: string,
-  description: string,
-  updated: string,
-}
