--- src/models/StateResponse.ts
@@ -1,12 +1,0 @@
-import { StateDetailsResponse } from '@/models/StateDetailsResponse'
-import { StateType } from '@/models/StateType'
-
-export type StateResponse = {
-  id: string,
-  type: Uppercase<StateType>,
-  message: string,
-  state_details: StateDetailsResponse | null,
-  data: Record<string, unknown>,
-  timestamp: string,
-  name: string,
-}
