--- src/models/WorkQueueFilterResponse.ts
@@ -1,4 +1,0 @@
-export type WorkQueueFilterResponse = {
-  tags: string[] | null,
-  deployment_ids: string[] | null,
-}
