+++ src/services/WorkspaceDeploymentsApi.ts
@@ -1,8 +1,9 @@
-import { DeploymentFlowRunCreate, DeploymentUpdate } from '@/models'
+import { DeploymentResponse } from '@/models/api/DeploymentResponse'
+import { FlowRunResponse } from '@/models/api/FlowRunResponse'
 import { Deployment } from '@/models/Deployment'
-import { DeploymentResponse } from '@/models/DeploymentResponse'
+import { DeploymentFlowRunCreate } from '@/models/DeploymentFlowRunCreate'
+import { DeploymentUpdate } from '@/models/DeploymentUpdate'
 import { FlowRun } from '@/models/FlowRun'
-import { FlowRunResponse } from '@/models/FlowRunResponse'
 import { mapper } from '@/services/Mapper'
 import { WorkspaceApi } from '@/services/WorkspaceApi'
 import { UnionFilters } from '@/types/UnionFilters'
