+++ src/services/WorkspaceFlowRunsApi.ts
@@ -1,8 +1,8 @@
 import { StateUpdate, TimelineNode } from '@/models'
+import { FlowRunGraphResponse } from '@/models/api/FlowRunGraphResponse'
+import { FlowRunHistoryResponse } from '@/models/api/FlowRunHistoryResponse'
+import { FlowRunResponse } from '@/models/api/FlowRunResponse'
 import { FlowRun } from '@/models/FlowRun'
-import { FlowRunGraphResponse } from '@/models/FlowRunGraphResponse'
-import { FlowRunHistoryResponse } from '@/models/FlowRunHistoryResponse'
-import { FlowRunResponse } from '@/models/FlowRunResponse'
 import { GraphNode } from '@/models/GraphNode'
 import { RunHistory } from '@/models/RunHistory'
 import { mapper } from '@/services/Mapper'
