+++ src/services/WorkspaceTaskRunsApi.ts
@@ -1,6 +1,6 @@
+import { TaskRunResponse } from '@/models/api/TaskRunResponse'
 import { StateUpdate } from '@/models/StateUpdate'
 import { TaskRun } from '@/models/TaskRun'
-import { TaskRunResponse } from '@/models/TaskRunResponse'
 import { mapper } from '@/services/Mapper'
 import { WorkspaceApi } from '@/services/WorkspaceApi'
 import { UnionFilters } from '@/types/UnionFilters'
