+++ src/types/index.ts
@@ -4,7 +4,7 @@ export * from './classes'
 export * from './ColorMode'
 export * from './css'
 export * from './dates'
-export * from './LogsRequestFilter'
+export * from '../models/api/LogsRequestFilter'
 export * from './reactivity'
 export * from './RouteGuard'
 export * from './SortOptionTypes'
